test if Imperator religion modifiers are correctly loaded

+++ ImperatorToCK3.UnitTests/CK3/Religions/HolySiteTests.cs
@@ -55,7 +55,7 @@ public class HolySiteTests {
 		titles.LoadTitles(titlesReader);
 		
 		var holySiteEffectMapper = new HolySiteEffectMapper("TestFiles/configurables/holy_site_effect_mappings.txt");
-		var imperatorEffects = new Dictionary<string, float> {
+		var imperatorEffects = new Dictionary<string, double> {
 			{"discipline", 0.2f}, // will be converted to knight_effectiveness_mult with factor of 10
 			{"unmapped_effect", 1f} // will be skipped
 		};
