+++ ImperatorToCK3/Imperator/Religions/Religion.cs
@@ -5,10 +5,10 @@ namespace ImperatorToCK3.Imperator.Religions;
 
 public sealed class Religion : IIdentifiable<string> {
 	public string Id { get; }
-	public IDictionary<string, float> Modifiers { get; }
+	public IDictionary<string, double> Modifiers { get; }
 
-	public Religion(string id, IDictionary<string, float> modifiers) {
+	public Religion(string id, IDictionary<string, double> modifiers) {
 		Id = id;
-		Modifiers = new Dictionary<string, float>(modifiers);
+		Modifiers = new Dictionary<string, double>(modifiers);
 	}
 }
